+++ tests/Zend/Session/Config/StandardConfigTest.php
@@ -10,20 +10,24 @@
 
 namespace ZendTest\Session;
 
-use Zend\Session\Configuration\SessionConfiguration;
+use Zend\Session\Config\StandardConfig;
 
 /**
  * @category   Zend
  * @package    Zend_Session
  * @subpackage UnitTests
  * @group      Zend_Session
- * @runTestsInSeparateProcesses
  */
-class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
+class StandardConfigTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var StandardConfig
+     */
+    protected $config;
+
     public function setUp()
     {
-        $this->config = new SessionConfiguration;
+        $this->config = new StandardConfig;
     }
 
     // session.save_path
@@ -34,29 +38,12 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->config->setSavePath(__DIR__ . '/foobarboguspath');
     }
 
-    public function testSavePathDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.save_path'), $this->config->getSavePath());
-    }
-
     public function testSavePathIsMutable()
     {
         $this->config->setSavePath(__DIR__);
         $this->assertEquals(__DIR__, $this->config->getSavePath());
     }
-
-    public function testSavePathAltersIniSetting()
-    {
-        $this->config->setSavePath(__DIR__);
-        $this->assertEquals(__DIR__, ini_get('session.save_path'));
-    }
-
     // session.name
-
-    public function testNameDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.name'), $this->config->getName());
-    }
 
     public function testNameIsMutable()
     {
@@ -64,18 +51,7 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('FOOBAR', $this->config->getName());
     }
 
-    public function testNameAltersIniSetting()
-    {
-        $this->config->setName('FOOBAR');
-        $this->assertEquals('FOOBAR', ini_get('session.name'));
-    }
-
     // session.save_handler
-
-    public function testSaveHandlerDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.save_handler'), $this->config->getSaveHandler(), var_export($this->config->toArray(), 1));
-    }
 
     public function testSaveHandlerIsMutable()
     {
@@ -83,35 +59,12 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('user', $this->config->getSaveHandler());
     }
 
-    public function testSaveHandlerAltersIniSetting()
-    {
-        $this->config->setSaveHandler('user');
-        $this->assertEquals('user', ini_get('session.save_handler'));
-    }
-
-    public function testSettingInvalidSaveHandlerRaisesException()
-    {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid save handler specified');
-        $this->config->setPhpSaveHandler('foobar_bogus');
-    }
-
     // session.gc_probability
-
-    public function testGcProbabilityDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.gc_probability'), $this->config->getGcProbability());
-    }
 
     public function testGcProbabilityIsMutable()
     {
         $this->config->setGcProbability(20);
         $this->assertEquals(20, $this->config->getGcProbability());
-    }
-
-    public function testGcProbabilityAltersIniSetting()
-    {
-        $this->config->setGcProbability(24);
-        $this->assertEquals(24, ini_get('session.gc_probability'));
     }
 
     public function testSettingInvalidGcProbabilityRaisesException()
@@ -134,21 +87,10 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
 
     // session.gc_divisor
 
-    public function testGcDivisorDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.gc_divisor'), $this->config->getGcDivisor());
-    }
-
     public function testGcDivisorIsMutable()
     {
         $this->config->setGcDivisor(20);
         $this->assertEquals(20, $this->config->getGcDivisor());
-    }
-
-    public function testGcDivisorAltersIniSetting()
-    {
-        $this->config->setGcDivisor(24);
-        $this->assertEquals(24, ini_get('session.gc_divisor'));
     }
 
     public function testSettingInvalidGcDivisorRaisesException()
@@ -165,21 +107,10 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
 
     // session.gc_maxlifetime
 
-    public function testGcMaxlifetimeDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.gc_maxlifetime'), $this->config->getGcMaxlifetime());
-    }
-
     public function testGcMaxlifetimeIsMutable()
     {
         $this->config->setGcMaxlifetime(20);
         $this->assertEquals(20, $this->config->getGcMaxlifetime());
-    }
-
-    public function testGcMaxlifetimeAltersIniSetting()
-    {
-        $this->config->setGcMaxlifetime(24);
-        $this->assertEquals(24, ini_get('session.gc_maxlifetime'));
     }
 
     public function testSettingInvalidGcMaxlifetimeRaisesException()
@@ -196,11 +127,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
 
     // session.serialize_handler
 
-    public function testSerializeHandlerDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.serialize_handler'), $this->config->getSerializeHandler());
-    }
-
     public function testSerializeHandlerIsMutable()
     {
         $value = extension_loaded('wddx') ? 'wddx' : 'php_binary';
@@ -208,25 +134,7 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($value, $this->config->getSerializeHandler());
     }
 
-    public function testSerializeHandlerAltersIniSetting()
-    {
-        $value = extension_loaded('wddx') ? 'wddx' : 'php_binary';
-        $this->config->setSerializeHandler($value);
-        $this->assertEquals($value, ini_get('session.serialize_handler'));
-    }
-
-    public function testSettingInvalidSerializeHandlerRaisesException()
-    {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid serialize handler specified');
-        $this->config->setSerializeHandler('foobar_bogus');
-    }
-
     // session.cookie_lifetime
-
-    public function testCookieLifetimeDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.cookie_lifetime'), $this->config->getCookieLifetime());
-    }
 
     public function testCookieLifetimeIsMutable()
     {
@@ -234,16 +142,10 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(20, $this->config->getCookieLifetime());
     }
 
-    public function testCookieLifetimeAltersIniSetting()
-    {
-        $this->config->setCookieLifetime(24);
-        $this->assertEquals(24, ini_get('session.cookie_lifetime'));
-    }
-
     public function testCookieLifetimeCanBeZero()
     {
         $this->config->setCookieLifetime(0);
-        $this->assertEquals(0, ini_get('session.cookie_lifetime'));
+        $this->assertEquals(0, $this->config->getCookieLifetime());
     }
 
     public function testSettingInvalidCookieLifetimeRaisesException()
@@ -260,21 +162,10 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
 
     // session.cookie_path
 
-    public function testCookiePathDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.cookie_path'), $this->config->getCookiePath());
-    }
-
     public function testCookiePathIsMutable()
     {
         $this->config->setCookiePath('/foo');
         $this->assertEquals('/foo', $this->config->getCookiePath());
-    }
-
-    public function testCookiePathAltersIniSetting()
-    {
-        $this->config->setCookiePath('/bar');
-        $this->assertEquals('/bar', ini_get('session.cookie_path'));
     }
 
     public function testSettingInvalidCookiePathRaisesException()
@@ -297,11 +188,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
 
     // session.cookie_domain
 
-    public function testCookieDomainDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.cookie_domain'), $this->config->getCookieDomain());
-    }
-
     public function testCookieDomainIsMutable()
     {
         $this->config->setCookieDomain('example.com');
@@ -312,12 +198,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
     {
         $this->config->setCookieDomain('');
         $this->assertEquals('', $this->config->getCookieDomain());
-    }
-
-    public function testCookieDomainAltersIniSetting()
-    {
-        $this->config->setCookieDomain('localhost');
-        $this->assertEquals('localhost', ini_get('session.cookie_domain'));
     }
 
     public function testSettingInvalidCookieDomainRaisesException()
@@ -334,111 +214,48 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
 
     // session.cookie_secure
 
-    public function testCookieSecureDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.cookie_secure'), $this->config->getCookieSecure());
-    }
-
     public function testCookieSecureIsMutable()
     {
-        $value = ini_get('session.cookie_secure') ? false : true;
-        $this->config->setCookieSecure($value);
-        $this->assertEquals($value, $this->config->getCookieSecure());
-    }
-
-    public function testCookieSecureAltersIniSetting()
-    {
-        $value = ini_get('session.cookie_secure') ? false : true;
-        $this->config->setCookieSecure($value);
-        $this->assertEquals($value, ini_get('session.cookie_secure'));
+        $this->config->setCookieSecure(true);
+        $this->assertEquals(true, $this->config->getCookieSecure());
     }
 
     // session.cookie_httponly
 
-    public function testCookieHttpOnlyDefaultsToIniSettings()
-    {
-        $this->assertSame((bool) ini_get('session.cookie_httponly'), $this->config->getCookieHttpOnly());
-    }
-
     public function testCookieHttpOnlyIsMutable()
     {
-        $value = ini_get('session.cookie_httponly') ? false : true;
-        $this->config->setCookieHttpOnly($value);
-        $this->assertEquals($value, $this->config->getCookieHttpOnly());
-    }
-
-    public function testCookieHttpOnlyAltersIniSetting()
-    {
-        $value = ini_get('session.cookie_httponly') ? false : true;
-        $this->config->setCookieHttpOnly($value);
-        $this->assertEquals($value, ini_get('session.cookie_httponly'));
+        $this->config->setCookieHttpOnly(true);
+        $this->assertEquals(true, $this->config->getCookieHttpOnly());
     }
 
     // session.use_cookies
 
-    public function testUseCookiesDefaultsToIniSettings()
-    {
-        $this->assertSame((bool) ini_get('session.use_cookies'), $this->config->getUseCookies());
-    }
-
     public function testUseCookiesIsMutable()
     {
-        $value = ini_get('session.use_cookies') ? false : true;
-        $this->config->setUseCookies($value);
-        $this->assertEquals($value, (bool) $this->config->getUseCookies());
-    }
-
-    public function testUseCookiesAltersIniSetting()
-    {
-        $value = ini_get('session.use_cookies') ? false : true;
-        $this->config->setUseCookies($value);
-        $this->assertEquals($value, (bool) ini_get('session.use_cookies'));
+        $this->config->setUseCookies(true);
+        $this->assertEquals(true, (bool) $this->config->getUseCookies());
     }
 
     // session.use_only_cookies
 
-    public function testUseOnlyCookiesDefaultsToIniSettings()
-    {
-        $this->assertSame((bool) ini_get('session.use_only_cookies'), $this->config->getUseOnlyCookies());
-    }
-
     public function testUseOnlyCookiesIsMutable()
     {
-        $value = ini_get('session.use_only_cookies') ? false : true;
-        $this->config->setOption('use_only_cookies', $value);
-        $this->assertEquals($value, (bool) $this->config->getOption('use_only_cookies'));
-    }
-
-    public function testUseOnlyCookiesAltersIniSetting()
-    {
-        $value = ini_get('session.use_only_cookies') ? false : true;
-        $this->config->setOption('use_only_cookies', $value);
-        $this->assertEquals($value, (bool) ini_get('session.use_only_cookies'));
+        $this->config->setUseOnlyCookies(true);
+        $this->assertEquals(true, (bool) $this->config->getUseOnlyCookies());
     }
 
     // session.referer_check
 
-    public function testRefererCheckDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.referer_check'), $this->config->getRefererCheck());
-    }
-
     public function testRefererCheckIsMutable()
     {
-        $this->config->setOption('referer_check', 'FOOBAR');
-        $this->assertEquals('FOOBAR', $this->config->getOption('referer_check'));
+        $this->config->setRefererCheck('FOOBAR');
+        $this->assertEquals('FOOBAR', $this->config->getRefererCheck());
     }
 
     public function testRefererCheckMayBeEmpty()
     {
-        $this->config->setOption('referer_check', '');
-        $this->assertEquals('', $this->config->getOption('referer_check'));
-    }
-
-    public function testRefererCheckAltersIniSetting()
-    {
-        $this->config->setOption('referer_check', 'BARBAZ');
-        $this->assertEquals('BARBAZ', ini_get('session.referer_check'));
+        $this->config->setRefererCheck('');
+        $this->assertEquals('', $this->config->getRefererCheck());
     }
 
     // session.entropy_file
@@ -455,29 +272,13 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->config->setEntropyFile(__DIR__);
     }
 
-    public function testEntropyFileDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.entropy_file'), $this->config->getEntropyFile());
-    }
-
     public function testEntropyFileIsMutable()
     {
         $this->config->setEntropyFile(__FILE__);
         $this->assertEquals(__FILE__, $this->config->getEntropyFile());
     }
 
-    public function testEntropyFileAltersIniSetting()
-    {
-        $this->config->setEntropyFile(__FILE__);
-        $this->assertEquals(__FILE__, ini_get('session.entropy_file'));
-    }
-
     // session.entropy_length
-
-    public function testEntropyLengthDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.entropy_length'), $this->config->getEntropyLength());
-    }
 
     public function testEntropyLengthIsMutable()
     {
@@ -485,16 +286,10 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(20, $this->config->getEntropyLength());
     }
 
-    public function testEntropyLengthAltersIniSetting()
-    {
-        $this->config->setEntropyLength(24);
-        $this->assertEquals(24, ini_get('session.entropy_length'));
-    }
-
     public function testEntropyLengthCanBeZero()
     {
         $this->config->setEntropyLength(0);
-        $this->assertEquals(0, ini_get('session.entropy_length'));
+        $this->assertEquals(0, $this->config->getEntropyLength());
     }
 
     public function testSettingInvalidEntropyLengthRaisesException()
@@ -521,11 +316,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function testCacheLimiterDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.cache_limiter'), $this->config->getCacheLimiter());
-    }
-
     /**
      * @dataProvider cacheLimiters
      */
@@ -535,27 +325,7 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($cacheLimiter, $this->config->getCacheLimiter());
     }
 
-    /**
-     * @dataProvider cacheLimiters
-     */
-    public function testCacheLimiterAltersIniSetting($cacheLimiter)
-    {
-        $this->config->setCacheLimiter($cacheLimiter);
-        $this->assertEquals($cacheLimiter, ini_get('session.cache_limiter'));
-    }
-
-    public function testSettingInvalidCacheLimiterRaisesException()
-    {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid cache limiter provided');
-        $this->config->setCacheLimiter('foobar_bogus');
-    }
-
     // session.cache_expire
-
-    public function testCacheExpireDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.cache_expire'), $this->config->getCacheExpire());
-    }
 
     public function testCacheExpireIsMutable()
     {
@@ -563,43 +333,26 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(20, $this->config->getCacheExpire());
     }
 
-    public function testCacheExpireAltersIniSetting()
-    {
-        $this->config->setCacheExpire(24);
-        $this->assertEquals(24, ini_get('session.cache_expire'));
-    }
-
     public function testSettingInvalidCacheExpireRaisesException()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid cache_expire; must be numeric');
+        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException',
+                                    'Invalid cache_expire; must be numeric');
         $this->config->setCacheExpire('foobar_bogus');
     }
 
     public function testSettingInvalidCacheExpireRaisesException2()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid cache_expire; must be a positive integer');
+        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException',
+                                    'Invalid cache_expire; must be a positive integer');
         $this->config->setCacheExpire(-1);
     }
 
     // session.use_trans_sid
 
-    public function testUseTransSidDefaultsToIniSettings()
-    {
-        $this->assertSame((bool) ini_get('session.use_trans_sid'), $this->config->getUseTransSid());
-    }
-
     public function testUseTransSidIsMutable()
     {
-        $value = ini_get('session.use_trans_sid') ? false : true;
-        $this->config->setOption('use_trans_sid', $value);
-        $this->assertEquals($value, (bool) $this->config->getOption('use_trans_sid'));
-    }
-
-    public function testUseTransSidAltersIniSetting()
-    {
-        $value = ini_get('session.use_trans_sid') ? false : true;
-        $this->config->setOption('use_trans_sid', $value);
-        $this->assertEquals($value, (bool) ini_get('session.use_trans_sid'));
+        $this->config->setUseTransSid(true);
+        $this->assertEquals(true, (bool) $this->config->getUseTransSid());
     }
 
     // session.hash_function
@@ -614,11 +367,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         return $provider;
     }
 
-    public function testHashFunctionDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.hash_function'), $this->config->getHashFunction());
-    }
-
     /**
      * @dataProvider hashFunctions
      */
@@ -626,21 +374,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
     {
         $this->config->setHashFunction($hashFunction);
         $this->assertEquals($hashFunction, $this->config->getHashFunction());
-    }
-
-    /**
-     * @dataProvider hashFunctions
-     */
-    public function testHashFunctionAltersIniSetting($hashFunction)
-    {
-        $this->config->setHashFunction($hashFunction);
-        $this->assertEquals($hashFunction, ini_get('session.hash_function'));
-    }
-
-    public function testSettingInvalidHashFunctionRaisesException()
-    {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid hash function provided');
-        $this->config->setHashFunction('foobar_bogus');
     }
 
     // session.hash_bits_per_character
@@ -654,11 +387,6 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function testHashBitsPerCharacterDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('session.hash_bits_per_character'), $this->config->getHashBitsPerCharacter());
-    }
-
     /**
      * @dataProvider hashBitsPerCharacters
      */
@@ -668,27 +396,14 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($hashBitsPerCharacter, $this->config->getHashBitsPerCharacter());
     }
 
-    /**
-     * @dataProvider hashBitsPerCharacters
-     */
-    public function testHashBitsPerCharacterAltersIniSetting($hashBitsPerCharacter)
-    {
-        $this->config->setHashBitsPerCharacter($hashBitsPerCharacter);
-        $this->assertEquals($hashBitsPerCharacter, ini_get('session.hash_bits_per_character'));
-    }
-
     public function testSettingInvalidHashBitsPerCharacterRaisesException()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid hash bits per character provided');
+        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException',
+                                    'Invalid hash bits per character provided');
         $this->config->setHashBitsPerCharacter('foobar_bogus');
     }
 
     // url_rewriter.tags
-
-    public function testUrlRewriterTagsDefaultsToIniSettings()
-    {
-        $this->assertSame(ini_get('url_rewriter.tags'), $this->config->getUrlRewriterTags());
-    }
 
     public function testUrlRewriterTagsIsMutable()
     {
@@ -696,26 +411,39 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('a=href,form=action', $this->config->getUrlRewriterTags());
     }
 
-    public function testUrlRewriterTagsAltersIniSetting()
-    {
-        $this->config->setUrlRewriterTags('a=href,fieldset=');
-        $this->assertEquals('a=href,fieldset=', ini_get('url_rewriter.tags'));
-    }
-
     // remember_me_seconds
-
-    public function testRememberMeSecondsDefaultsToTwoWeeks()
-    {
-        $this->assertEquals(1209600, $this->config->getRememberMeSeconds());
-    }
 
     public function testRememberMeSecondsIsMutable()
     {
-        $this->config->setRememberMeSeconds(604800);
-        $this->assertEquals(604800, $this->config->getRememberMeSeconds());
+        $this->config->setRememberMeSeconds(20);
+        $this->assertEquals(20, $this->config->getRememberMeSeconds());
+    }
+
+    public function testSettingInvalidRememberMeSecondsRaisesException()
+    {
+        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException',
+                                    'Invalid remember_me_seconds; must be numeric');
+        $this->config->setRememberMeSeconds('foobar_bogus');
+    }
+
+    public function testSettingInvalidRememberMeSecondsRaisesException2()
+    {
+        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException',
+                                    'Invalid remember_me_seconds; must be a positive integer');
+        $this->config->setRememberMeSeconds(-1);
     }
 
     // setOptions
+
+    /**
+     * @dataProvider optionsProvider
+     */
+    public function testSetOptionsTranslatesUnderscoreSeparatedKeys($option, $getter, $value)
+    {
+        $options = array($option => $value);
+        $this->config->setOptions($options);
+        $this->assertSame($value, $this->config->$getter());
+    }
 
     public function optionsProvider()
     {
@@ -732,7 +460,7 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
             ),
             array(
                 'save_handler',
-                'getOption',
+                'getSaveHandler',
                 'user',
             ),
             array(
@@ -827,7 +555,7 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
             ),
             array(
                 'hash_bits_per_character',
-                'getOption',
+                'getHashBitsPerCharacter',
                 5,
             ),
             array(
@@ -836,19 +564,5 @@ class SessionConfigurationTest extends \PHPUnit_Framework_TestCase
                 'a=href',
             ),
         );
-    }
-
-    /**
-     * @dataProvider optionsProvider
-     */
-    public function testSetOptionsTranslatesUnderscoreSeparatedKeys($option, $getter, $value)
-    {
-        $options = array($option => $value);
-        $this->config->setOptions($options);
-        if ('getOption' == $getter) {
-            $this->assertSame($value, $this->config->getOption($option));
-        } else {
-            $this->assertSame($value, $this->config->$getter());
-        }
     }
 }
